initial commit with minor changes and docs
