+++ requirements.txt
@@ -1,1 +1,1 @@
-boto==2.49.0
+boto3==1.34.109
